+++ src/pages/ReservationPage/Reservation.css
@@ -109,7 +109,7 @@
 }
 
 
-.nextBtn {
+.nextBtnOnly {
   text-align: center;
   margin: 50px 1rem;
   padding: 1.5rem 2.5rem;
