+++ src/pages/ReservationPage/RsvnRoom.css
@@ -23,25 +23,12 @@
 
 /* 사이즈 공통 */
 .rsvnWrap2 {
-  position: relative;
-  /* max-width: 50rem; */
-  margin: 0 auto 100px;
-  overflow: hidden;
-  
-  background: pink;
-}
-
-.tabBtn {
-  display: inline-block;
-  padding: 10px;
-  border: 1px solid black;
-  cursor: pointer;
-}
-
-.tabBtnSelected { 
-  background: pink;
-}
-
-.roomLayout {
   display: flex;
+  position: relative;
+  justify-content: space-between;
+  max-width: 80rem;
+  margin: 0 auto 100px;
+  /* overflow: hidden; */
+  
+  background: moccasin;
 }
